# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Go+Vue实现的openLDAP后台管理项目",
        "title": "Go Ldap Admin",
        "termsOfService": "https://github.com/eryajf/go-ldap-admin",
        "contact": {
            "name": "项目作者：二丫讲梵 、 swagger作者：南宫乘风",
            "url": "https://github.com/eryajf/go-ldap-admin",
            "email": "https://github.com/eryajf/go-ldap-admin"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8888",
    "basePath": "/api",
    "paths": {
...
        "/group/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加分组记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "添加分组记录",
                "parameters": [
                    {
                        "description": "添加用户记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/adduser": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "description": "添加用户记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupAddUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除分组记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "删除分组记录",
                "parameters": [
                    {
                        "description": "删除用户记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取分组记录列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "获取分组记录列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/removeuser": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "移除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "移除用户",
                "parameters": [
                    {
                        "description": "移除用户记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupRemoveUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/syncDingTalkDepts": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "同步钉钉部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "同步钉钉部门信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/syncFeiShuDepts": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "同步飞书部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "同步飞书部门信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/syncOpenLdapDepts": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "同步原ldap部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "同步原ldap部门信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/syncSqlGroups": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "同步Sql中的分组信息到ldap",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "同步Sql中的分组信息到ldap",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/syncWeComDepts": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "同步企业微信部门信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "同步企业微信部门信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/tree": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取分组接口树",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "获取分组接口树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新分组记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "更新分组记录",
                "parameters": [
                    {
                        "description": "更新用户记录的结构体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/useringroup": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取分组内用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "获取分组内用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickname",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
        "/group/usernoingroup": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "不在分组的用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分组管理"
                ],
                "summary": "不在分组的用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickname",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.GroupAddReq": {
            "type": "object",
            "required": [
                "groupName",
                "groupType"
            ],
            "properties": {
                "groupName": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1
                },
                "groupType": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "parentId": {
                    "description": "父级Id 大于等于0 必填",
                    "type": "integer",
                    "minimum": 0
                },
                "remark": {
                    "description": "分组的中文描述",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 0
                }
            }
        },
        "request.GroupAddUserReq": {
            "type": "object",
            "required": [
                "groupId",
                "userIds"
            ],
            "properties": {
                "groupId": {
                    "type": "integer"
                },
                "userIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "request.GroupDeleteReq": {
            "type": "object",
            "required": [
                "groupIds"
            ],
            "properties": {
                "groupIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "request.GroupRemoveUserReq": {
            "type": "object",
            "required": [
                "groupId",
                "userIds"
            ],
            "properties": {
                "groupId": {
                    "type": "integer"
                },
                "userIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "request.GroupUpdateReq": {
            "type": "object",
            "required": [
                "groupName",
                "id"
            ],
            "properties": {
                "groupName": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "description": "分组的中文描述",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 0
                }
            }
        },
...
        "response.ResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}